Have chashdatum.len be unsigned int to preserve binary compatibility with master.
This requires casts from size_t to unsigned int for 64-bit builds
where this field is set from strlen (43 occurrences) and DBT.size
(2 occurrences).

+++ src/data-types/mail_cache_db.c
@@ -348,7 +348,7 @@ int mail_cache_db_clean_up(struct mail_cache_db * cache_db,
     chashdatum hash_data;
     
     hash_key.data = db_key.data;
-    hash_key.len = db_key.size;
+    hash_key.len = (unsigned int)db_key.size;
 
     r = chash_get(exist, &hash_key, &hash_data);
     if (r < 0) {
@@ -475,7 +475,7 @@ int mail_cache_db_get_keys(struct mail_cache_db * cache_db,
     chashdatum hash_data;
     
     hash_key.data = db_key.data;
-    hash_key.len = db_key.size;
+    hash_key.len = (unsigned int)db_key.size;
     hash_data.data = NULL;
     hash_data.len = 0;
