+++ src/engine/mailprivacy_gnupg.c
@@ -3003,9 +3003,9 @@ int mailprivacy_gnupg_set_encryption_id(struct mailprivacy * privacy,
   }
   
   key.data = buf;
-  key.len = strlen(buf) + 1;
+  key.len = (unsigned int)strlen(buf) + 1;
   value.data = passphrase;
-  value.len = strlen(passphrase) + 1;
+  value.len = (unsigned int)strlen(passphrase) + 1;
   
   r = chash_set(passphrase_hash, &key, &value, NULL);
   if (r < 0) {
@@ -3034,7 +3034,7 @@ static char * get_passphrase(struct mailprivacy * privacy,
     return NULL;
   
   key.data = buf;
-  key.len = strlen(buf) + 1;
+  key.len = (unsigned int)strlen(buf) + 1;
   
   r = chash_get(passphrase_hash, &key, &value);
   if (r < 0)
